Need to test on linux machine.
Former-commit-id: 9549f62f3d8cafe1022a990e5f302ba2be4826fc

+++ Project3/src/ComparatorEdge.java
@@ -6,7 +6,7 @@ public class ComparatorEdge implements Comparator {
 		Edge e1 = (Edge) o1;
 		Edge e2 = (Edge) o2;
 
-		return Integer.compare(e1.getWeight(), e2.getWeight());
+		return Long.compare(e1.getWeight(), e2.getWeight());
 	}
 
 }
